Restructure the module folder
Move files into documents and helper folders.
Update all the import statement.
